pyproject.toml
[project]
name = "vr_dash"
version = "1.2.0"
description = ""
authors = [{name = "Matthias Hauth", email ="matthias.hauth@deltares.nl"}]
maintainers = [
]
readme = "README.md"
keywords = ["civil-engineering", "dikes", "flood-defence", "deltares", "dashboard", "vrtool"]
classifiers = [
    'Programming Language :: Python :: 3',
    'Natural Language :: English',
    'Intended Audience :: Science/Research',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Topic :: Scientific/Engineering',
    'Topic :: Scientific/Engineering :: GIS',
    'Topic :: Scientific/Engineering :: Mathematics',
    'Topic :: Scientific/Engineering :: Physics',
    'Topic :: Scientific/Engineering :: Hydrology',
]
requires-python = "==3.12"
dependencies = ["vrtool @ https://github.com/Deltares/Veiligheidsrendement/releases/download/v1.2.0/vrtool-1.2.0-py3-none-any.whl", "geopandas", "numpy", "tabulate", "diskcache", "multiprocess", "plotly", "dash", "dash-ag-grid", "dash-iconify", "dash-mantine-components", "dash-bootstrap-components", "openturns"]

[project.urls]
homepage = "https://deltares-research.github.io/VrtoolDocumentation/"
documentation = "https://deltares-research.github.io/VrtoolDocumentation/"
repository = "https://github.com/Deltares/VrtoolDashboard"
issues = "https://github.com/Deltares/veiligheidsrendement/issues"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pixi.pypi-dependencies]
vr_dash = { path = ".", editable = true }
build = "*"

[tool.pixi.workspace]
channels = ["conda-forge"]
platforms = ["linux-64","osx-arm64", "osx-64", "win-64"]

[tool.pixi.dependencies]
python = "==3.12"
geopandas = ">=1.1.1"
numpy = ">=1.26.4"
tabulate = ">=0.9.0"
diskcache = ">=5.6.3"
plotly = ">=5.15,<6"
dash = ">=2.17.0,<3"
dash-ag-grid = ">=2.4.0"
dash-iconify = ">=0.1.2"
dash-mantine-components = ">=0.14.0"
dash-bootstrap-components = ">=1.4.1,<2.0.0"
openturns = ">=1.26,<2"
multiprocess = ">=0.70.18,<0.71"

[tool.pixi.feature.test.dependencies]
pytest = ">=7.2.1"
pytest-cov = ">=3.0.0"
coverage = ">=6.4.4"
teamcity-messages = ">=1.32"
pytest-xdist = ">=3.3.1"
pytest-profiling = ">=1.7.0"

[tool.pixi.feature.dev.dependencies]
black = ">=22.8.0"
isort = ">=5.10.1"
commitizen = ">=4.10.1,<5"
urllib3 = "==1.26.15"

[dependency-groups]
dev = ["build>=1.3.0,<2"]



[tool.pixi.feature.dev.tasks]
isort = "isort"
black = "black"
build = "python -m build --outdir dist/"


[tool.pixi.feature.py312.dependencies]
python = "3.12.*"


[tool.pixi.environments]
default = { solve-group = "default" }
test = { features = ["test"], solve-group = "default" }
dev = { features = ["dev", "test"], solve-group = "default" }
py312 = ["py312", "default"]


[tool.black]
line-length = 88
target-version = ['py312']
exclude = '''
(
  /(
      \.git         # exclude a few common directories in the
    | \.github          # root of the project
    | \.env
    | \.venv
    | \.pixi
    | \.config
    | .pytest_cache
    | .scannerwork
    | .vscode
    | site
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
skip = [".git", ".github", ".pixi", ".env", ".venv", ".config", ".pytest_cache", ".scannerwork", ".vscode", "site"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "externals: marks tests that require of external binaries (deselect with '-m \"not externals\"')",
]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.2.0"
tag_format = "v$major.$minor.$patch"
changelog_file = "docs/CHANGELOG.md"
update_changelog_on_bump = true
version_files = [
    "src/__init__.py",
    "pyproject.toml:version",
    "pixi.toml:version"]

